% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scat_lin_cor.R
\name{scat.lin.cor}
\alias{scat.lin.cor}
\title{Scatter plot with correlation and optional regression line}
\usage{
scat.lin.cor(
  data,
  x,
  y,
  xlab = x,
  ylab = y,
  alpha = 0.8,
  reg = T,
  ci = T,
  refline = F,
  BF = F,
  pval = F,
  method = "pearson"
)
}
\arguments{
\item{data}{A data frame containing the variables to be plotted.}

\item{x}{A string specifying the name of the variable on the x-axis.}

\item{y}{A string specifying the name of the variable on the y-axis.}

\item{xlab}{(optional) A string for the x-axis label. Defaults to the value of \code{x}.}

\item{ylab}{(optional) A string for the y-axis label. Defaults to the value of \code{y}.}

\item{alpha}{(optional) A numeric value for the transparency of points, between 0 and 1. Default is 0.8.}

\item{reg}{(optional) A logical indicating whether to fit a regression line. Default is TRUE.}

\item{ci}{(optional) A logical indicating whether to display a confidence interval around the regression line. Default is TRUE.}

\item{refline}{(optional) A logical indicating whether to add a reference line (y = x). Default is FALSE.}

\item{BF}{(optional) A logical indicating whether to display the Bayes Factor. Default is FALSE.}

\item{pval}{(optional) A logical indicating whether to display the p-value. Default is FALSE.}

\item{method}{(optional) A string specifying the correlation method to be used. Options are "pearson", "spearman", and "kendall". Default is "pearson".}
}
\value{
A \code{ggplot} object representing the scatter plot.
}
\description{
Quick scatter plot between two variables with regression line, correlation coefficient title, and optional reference lines. Designed for parameter recovery, model validation etc.
The function is similar to \code{scatr::scat} with more options but no univariate density plots.
}
\details{
The function computes the correlation between the variables \code{x} and \code{y} using the specified method
and displays the correlation coefficient in the plot title. If \code{BF} is TRUE, the function calculates
and displays the Bayes Factor for Pearson correlation (currently not supported for Spearman or Kendall).
If \code{pval} is TRUE, the p-value from the correlation test is also displayed. The function
adds a regression line, confidence interval, and optionally adds a reference line for comparison.
}
\examples{
# Example usage:
data(mtcars)
scat.lin.cor(mtcars, x = "mpg", y = "wt", reg = TRUE, ci = TRUE, pval = TRUE)

}
