% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wd_here.R
\name{wd.here}
\alias{wd.here}
\title{Set Working Directory to the Script's Location (requires RStudio but not RProj)}
\usage{
wd.here()
}
\value{
The new absolute working directory (path to current script), as a character string.
}
\description{
This function sets the working directory to the directory containing the currently active R script in RStudio.
It mimics MATLAB's behavior of automatically using the script's directory as the working directory when running the script.
}
\details{
This function may be particularly useful for users transitioning from MATLAB to R. The working directory in MATLAB is automatically set to the script's location.
Unlike \code{here::here()}, this function does not depend on the script being within an R project (\code{.RProj}) structure but instead relies on RStudio's API to identify the script's directory.
}
\examples{
# To set the working directory to the script's location, simply call:
wd.here()

}
\seealso{
\code{\link[here]{here}}, \code{\link[base]{setwd}}
}
